x86: type __ftol-style consumed arg as f64, not floatish
An x87 stack-consuming helper's argument was typed floatish, which unifies
across call sites to the first concrete width. A helper (e.g. __ftol) called
with both a float and a double therefore narrowed the double argument to f32
-- `__ftol((f32) _dbl)` for `(unsigned)dbl` -- silently losing precision. Type
it as the widest float so float args promote losslessly instead. Unblocks the
float-conversions /O2 fixture; int-float's __ftol call now shows the lossless
(f64) promote.

+++ tests/end_to_end/int-float/msvc-o2-out.c
@@ -1,4 +1,4 @@
-s32 __ftol(f32);                                    /* extern */
+s32 __ftol(f64);                                    /* extern */
 extern f32 _globalf;
 extern s32 _globali;
 static f32 _real_40a00000 = 5.0f;                   /* const */
@@ -7,7 +7,7 @@ f32 test(f32 arg0, s32 arg1, f32 arg2, s32 arg3) {
     s32 sp0;
     s32 sp4;
 
-    _globali = __ftol(arg0);
+    _globali = __ftol((f64) arg0);
     sp4 = 0;
     _globalf = (f32) arg1;
     sp0 = arg3 + 3;
